+++ src/Files.App/Data/Models/ItemViewModel.cs
@@ -38,7 +38,7 @@ using static Files.Backend.Helpers.NativeFindStorageItemHelper;
 using DispatcherQueue = Microsoft.UI.Dispatching.DispatcherQueue;
 using FileAttributes = System.IO.FileAttributes;
 
-namespace Files.App.ViewModels
+namespace Files.App.Data.Models
 {
 	public sealed class ItemViewModel : ObservableObject, IDisposable
 	{
