+++ src/Files.App/Data/Models/NetworkDrivesViewModel.cs
@@ -5,7 +5,7 @@ using Files.App.Data.Items;
 using Files.Backend.Services;
 using Files.Sdk.Storage.LocatableStorage;
 
-namespace Files.App.ViewModels
+namespace Files.App.Data.Models
 {
 	public class NetworkDrivesViewModel : ObservableObject
 	{
