+++ src/Files.App/GlobalUsings.cs
@@ -31,6 +31,7 @@ global using global::Files.App.Data.Parameters;
 global using global::Files.App.Interacts;
 global using global::Files.App.UserControls;
 global using global::Files.App.ViewModels;
+global using global::Files.App.ViewModels.UserControls;
 global using global::Files.App.Views;
 global using global::Files.App.Views.LayoutModes;
 global using global::Files.App.Views.Shells;
